handle packagify changes as modified
