+++ Tests/schemas/integration.yml
@@ -2,24 +2,24 @@ type: map
 mapping:
   commonfields:
     type: map
-    required: yes
+    required: true
     mapping:
       id:
         type: str
-        required: yes
+        required: true
       version:
         type: int
-        required: yes
+        required: true
       sortvalues:
         type: seq
         sequence:
         - include: sortvalue_schema
   name:
     type: str
-    required: yes
+    required: true
   display:
       type: str
-      required: yes
+      required: true
   deprecated:
     type: bool
   beta:
@@ -31,7 +31,7 @@ mapping:
     'Vulnerability Management', 'Case Management',
     'Forensics & Malware Analysis', 'IT Services', 'Data Enrichment & Threat Intelligence',
      'Authentication', 'Database', 'Deception', 'Email Gateway']
-    required: yes
+    required: true
   fromversion:
     type: str
   toversion:
@@ -40,25 +40,25 @@ mapping:
     type: str
   description:
     type: str
-    required: yes
+    required: true
   detaileddescription:
     type: str
   configuration:
     type: seq
-    required: yes
+    required: true
     sequence:
     - include: configuration_schema
   script:
     type: map
-    required: yes
+    required: true
     mapping:
       script:
         type: str
-        required: yes
+        required: true
       type:
         type: str
         enum: ['javascript', 'python']
-        required: yes
+        required: true
       dockerimage:
         type: str
       alt_dockerimages:
@@ -98,19 +98,19 @@ schema;sortvalue_schema:
 # integration->configuration
 schema;configuration_schema:
   type: map
-  required: yes
+  required: true
   mapping:
     display:
       type: str
-      required: yes
+      required: true
     defaultvalue:
       type: any
     name:
       type: str
-      required: yes
+      required: true
     type:
       type: int
-      required: yes
+      required: true
     required:
       type: bool
     hidden:
@@ -126,12 +126,12 @@ schema;command_schema:
   mapping:
     name:
       type: str
-      required: yes
+      required: true
     execution:
       type: bool
     description:
       type: str
-      required: yes
+      required: true
     deprecated:
       type: bool
     system:
@@ -157,14 +157,14 @@ schema;argument_schema:
   mapping:
     name:
       type: str
-      required: yes
+      required: true
     required:
       type: bool
     default:
       type: bool
     description:
       type: str
-      required: yes
+      required: true
     auto:
       type: str
       enum: ['PREDEFINED']
@@ -192,7 +192,7 @@ schema;output_schema:
       type: str
     description:
       type: str
-      required: yes
+      required: true
     important:
       type: bool
     importantDescription:
@@ -206,9 +206,9 @@ schema;important_schema:
   mapping:
     contextPath:
       type: str
-      required: yes
+      required: true
     description:
       type: str
-      required: yes
+      required: true
     related:
       type: str
